requirements.txt
beautifulsoup4==4.6.0
bleach==1.4
boto==2.45.0
django-annoying==0.8.0
django-compressor==1.4
django-gravatar==0.1.0
django-mobile==0.4.0
django-registration==2.0
django-storages==1.4.1
dkimpy==0.6.1
dnspython==1.15.0
google-api-python-client==1.6.1
html2text==2014.9.25
imapclient==2.0.0
ipaddr==2.1.11
jsonfield==1.0.3
lamson==1.3.4
markdown2==2.3.0
mysql-python==1.2.5
pycrypto==2.6.1
pydns==2.3.6
pyspf==2.0.1
python-gflags==3.1.0
python-gnupg==0.4.0
pytz==2016.7
requests==2.18.4
south==1.0.1
typing==3.6.6